Data and format
It's taking the name from the console.

It's not saving the date anymore, since it's possible to take it from the startTime.

It's saving start and final time in a Date format.

There are two new boolean variables to save if the user had high or low heart rate in the experience.

It's saving the time of each heartbeat with the heart rate.

It's only saving the final date when the watch is disconnect. If the application is forced to close, it checks if there is a final time, if not it saves the current time.